% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_interval_Bayes_factor.R
\name{Beta_intervalBayesfactor}
\alias{Beta_intervalBayesfactor}
\title{Beta Interval Bayes Factor}
\usage{
Beta_intervalBayesfactor(
  n1,
  n2,
  lowerH0value = 0,
  upperH0value = 0.5,
  a0 = 1,
  b0 = 1
)
}
\arguments{
\item{n1}{Number of values for variable 1}

\item{n2}{Number of values for variable 2}

\item{lowerH0value}{(Optional) Lower limit of interval estimate for the null hypothesis (default is 0)}

\item{upperH0value}{(Optional) Upper limit of interval estimate for the null hypothesis (default is 0.5)}

\item{a0}{(Optional) Shape parameter alpha for the prior beta distribution (default is 1)}

\item{b0}{(Optional) Shape parameter beta for the prior beta distribution (default is 1)}
}
\value{
\item{H0_range}{Interval estimate of the null hypothesis -- either the default of [0, 0.5] or the value specified by user}

\item{H0_prior}{Prior probability of the null interval}

\item{H1_prior}{Prior probability of the alternative interval}

\item{H0_posterior}{Posterior probability of the null interval}

\item{H1_posterior}{Posterior probability of the alternative interval}

\item{Direction}{Statement describing the numerator and denominator of Bayes Factor estimate}

\item{BF}{Bayes Factor estimate}
}
\description{
Beta Interval Bayes Factor
}
