% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_point_Bayes_factor.R
\name{Beta_pointBayesfactor}
\alias{Beta_pointBayesfactor}
\title{Beta Point Bayes Factor}
\usage{
Beta_pointBayesfactor(n1, n2, ptH0 = 0.5, a0 = 1, b0 = 1)
}
\arguments{
\item{n1}{Number of values for variable 1}

\item{n2}{Number of values for variable 2}

\item{ptH0}{(Optional) Point estimate for the null hypothesis (default is 0.5)}

\item{a0}{(Optional) Shape parameter alpha for the prior beta distribution (default is 1)}

\item{b0}{(Optional) Shape parameter beta for the prior beta distribution (default is 1)}
}
\value{
\item{H0_point}{Value of the point null hypothesis -- either the default of 0.5 or the value specified by user}

\item{Direction}{Statement describing the numerator and denominator of Bayes Factor estimate}

\item{BF}{Bayes Factor estimate}
}
\description{
Beta Point Bayes Factor
}
