% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_bayesian_contrasts.R
\name{dfba_bayesian_contrasts}
\alias{dfba_bayesian_contrasts}
\title{Bayesian Contrasts}
\usage{
dfba_bayesian_contrasts(
  wts,
  n1v,
  n2v,
  priora0,
  priorb0,
  prob_interval = 0.95,
  samples = 10000
)
}
\arguments{
\item{wts}{A set of weights for contrast. Contrast weights must sum to 1.}

\item{n1v}{Number of successes in each condition}

\item{n2v}{Number of failures in each condition}

\item{priora0}{(Optional) Set of prior values for the beta shape parameter alpha for each condition (default is a = 1 for each condition)}

\item{priorb0}{(Optional) Set of prior values for the beta shape parameter beta for each condition (default is a = 1 for each condition)}

\item{prob_interval}{Desired probability for equal-tail interval estimate on the contrast (default is 0.95)}

\item{samples}{The desired number of Monte Carlo samples (default is 10000)}
}
\value{
A list containing the following components:

\item{contrast_Weights}{Contrast weights from user input}

\item{contrast_mean}{Mean of the contrast}

\item{contrast_variance}{Variance of the contrast}

\item{samples}{The number of Monte Carlo samples (default is 10000)}

\item{sampling_contrast_mean}{Mean of the contrast samples}

\item{sampling_contrast_variance}{Variance of the contrast samples}

\item{equal_tail_interval}{Limits of the posterior equal-tail interval estimate on the contrast with probability defined by \code{prob_interval}}

\item{posterior_of_positive}{Posterior probability that the contrast is positive}

\item{prior_of_positive}{Prior probability that the contrast is positive}

\item{BF_positive_contrast}{Bayes Factor estimate in favor of the contrast being positive}

\item{BF_negative_contrast}{Bayes Factor estimate in favor of the contrast being negative}
}
\description{
Compute contrasts for binomial data from multiple conditions
}
\details{
These are the details.
}
\examples{
wts <- c(1/3,1/3,1/3,-.5,-.5)
n1v <- c(8,7,10,3,6) # Successes per condition
n2v <- c(4,5,2,9,6)  # Failures per condition
priora0 <- c(1,1,1,1,1)
priorb0 <- c(1,1,1,1,1)
dfba_bayesian_contrasts(wts, n1v, n2v, priora0, priorb0)
}
