% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_beta_contrast.R
\name{show,dfba_beta_contrast_out-method}
\alias{show,dfba_beta_contrast_out-method}
\alias{plot,dfba_beta_contrast_out-method}
\title{Formats for Beta Contrasts}
\usage{
\S4method{show}{dfba_beta_contrast_out}(object)

\S4method{plot}{dfba_beta_contrast_out}(x)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_beta_contrast_out}}}

\item{x}{object An object of class \code{\linkS4class{dfba_beta_contrast_out}}}
}
\description{
Formats for Beta Contrasts
}
