% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_beta_descriptive.R
\name{show,dfba_beta_descriptive_out-method}
\alias{show,dfba_beta_descriptive_out-method}
\alias{plot,dfba_beta_descriptive_out-method}
\title{Formats for Beta Descriptive}
\usage{
\S4method{show}{dfba_beta_descriptive_out}(object)

\S4method{plot}{dfba_beta_descriptive_out}(x)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_beta_descriptive_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_beta_descriptive_out}}}
}
\value{
No return value, called for side effect. Objects of class \code{\linkS4class{dfba_beta_descriptive_out}} are printed.

No return value, called for side effect. Method produces a plot of class \code{\linkS4class{dfba_beta_descriptive_out}}
}
\description{
Formats for Beta Descriptive
}
\keyword{internal}
