% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_binomial.R
\name{show,dfba_binomial_out-method}
\alias{show,dfba_binomial_out-method}
\alias{plot,dfba_binomial_out-method}
\title{Formats for Bayesian Binomial Test}
\usage{
\S4method{show}{dfba_binomial_out}(object)

\S4method{plot}{dfba_binomial_out}(x, plot.prior = TRUE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_binomial_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_binomial_out}}}

\item{plot.prior}{Show prior distribution (default = TRUE)}
}
\description{
Formats for Bayesian Binomial Test
}
\keyword{internal}
