% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_classes.R
\docType{class}
\name{dfba_gamma_out-class}
\alias{dfba_gamma_out-class}
\alias{dfba_bivariate_concordance_out-class}
\alias{dfba_bivariate_concordance_star_out-class}
\alias{dfba_mann_whitney_small_out-class}
\alias{dfba_mann_whitney_large_out-class}
\alias{dfba_wilcoxon_small_out-class}
\alias{dfba_wilcoxon_large_out-class}
\alias{dfba_t_power_out-class}
\alias{dfba_power_curve_out-class}
\alias{dfba_point_BF_out-class}
\alias{dfba_interval_BF_out-class}
\alias{dfba_beta_contrast_out-class}
\alias{dfba_sim_data_out-class}
\alias{dfba_mcnemar_out-class}
\alias{dfba_median_test_out-class}
\alias{dfba_beta_descriptive_out-class}
\alias{dfba_sign_test_out-class}
\alias{dfba_binomial_out-class}
\title{Classes for DFBA}
\description{
Classes for DFBA
}
