% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_gamma.R
\name{dfba_gamma}
\alias{dfba_gamma}
\title{Goodman-Kruskal Gamma
This function takes either:
two vectors of equal length OR
a cross-tabulation of values
and returns:
  the Goodman-Kruskal Gamma Statistic
  the observed concordance statistic p
  the concordance parameter Phi
  shape parameters a and b (alpha and beta) on the
    beta distribution that describes Phi
  Highest Density Interval (HDI) limits on Phi}
\usage{
dfba_gamma(
  x,
  y = NULL,
  quantiles_x = NULL,
  quantiles_y = NULL,
  a.prior = 1,
  b.prior = 1,
  hdi.width = 0.95
)
}
\arguments{
\item{x}{vector of x variable values}

\item{y}{vector of y variable values}

\item{quantiles_x}{(Optional) If x is numeric, desired number of quantiles for the x variable}

\item{quantiles_y}{(Optional) If y is numeric  numeric, desired number of quantiles for the y variable}

\item{a.prior}{shape parameter a of the prior beta distribution}

\item{b.prior}{shape parameter b of the prior beta distribution}

\item{hdi.width}{Desired width of the highest density interval (HDI) of the posterior distribution (default is 95\%)}
}
\value{
A list containing the following components:

\item{Tau}{Nonparametric Tau-a correlation}

\item{sample_p}{Sample concordance proportion}

\item{nc}{Number of concordant (x, y) pairs}

\item{nd}{Number of discordant (x, y) pairs}

\item{post.median}{Median of posterior distribution on phi}

\item{post.hdi.lower}{lower limit of the HDI with width specified by hdi.width}

\item{post.hdi.upper}{upper limit of the HDI with width specified by hdi.width}
}
\description{
Goodman-Kruskal Gamma
This function takes either:
two vectors of equal length OR
a cross-tabulation of values
and returns:
  the Goodman-Kruskal Gamma Statistic
  the observed concordance statistic p
  the concordance parameter Phi
  shape parameters a and b (alpha and beta) on the
    beta distribution that describes Phi
  Highest Density Interval (HDI) limits on Phi
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
