% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_mann_whitney.R
\name{dfba_mann_whitney}
\alias{dfba_mann_whitney}
\title{Independent Samples Test (Mann Whitney U)}
\usage{
dfba_mann_whitney(
  E,
  C,
  a0 = 1,
  b0 = 1,
  prob_interval = 0.95,
  samples = 30000,
  method = ""
)
}
\arguments{
\item{E}{Data for independent sample 1 ("Experimental")}

\item{C}{Data for independent sample 2 ("Control")}

\item{a0}{shape parameter alpha of the prior beta distribution}

\item{b0}{shape parameter beta of the prior beta distribution}

\item{prob_interval}{Desired width of the highest density interval (HDI) of the posterior distribution (default is 95\%)}

\item{samples}{The number of desired Markov-Chain samples (default is 30000)}

\item{method}{(Optional) The method option is either "small" or "large". The "small" algorithm is based on a discrete Monte Carlo solution for cases where n is typically less than 20. The "large" algorithm is based on beta approximation model for the posterior distribution for the omega_E parameter. This approximation is reasonable when n > 19. Regardless of n the user can stipulate which method that they desire. When the method option is  omitted the program selects the appropriate procedure.}
}
\value{
A list containing the following components:

\item{Tau}{Nonparametric Tau-a correlation}

\item{sample_p}{Sample concordance proportion}

\item{nc}{Number of concordant (x, y) pairs}

\item{nd}{Number of discordant (x, y) pairs}

\item{post.median}{Median of posterior distribution on phi}

\item{post.eti.lower}{lower limit of the equal-tail interval with width specified by interval.width}

\item{post.eti.upper}{upper limit of the equal-tail interval with width specified by interval.width}
}
\description{
The Mann-Whitney U statistic is based on independent samples that are
denoted as E and C. prior_vec has two components that are the shape
parameters of the prior beta distribution. The default prior is the uniform
prior, but the user can stipulate with the components of prior_vec, the
desired prior distribution. The prob_interval is value for the
probability interval estimates. Samples is the number of Monte Carlo samples
used for the small n analysis.
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
