% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_mann_whitney_small.R
\name{show,dfba_mann_whitney_small_out-method}
\alias{show,dfba_mann_whitney_small_out-method}
\alias{plot,dfba_mann_whitney_small_out-method}
\title{Formats for small-n Mann Whitney}
\usage{
\S4method{show}{dfba_mann_whitney_small_out}(object)

\S4method{plot}{dfba_mann_whitney_small_out}(x, plot.prior = TRUE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_mann_whitney_small_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_mann_whitney_small_out}}}

\item{plot.prior}{Show prior distribution (default = TRUE)}
}
\description{
Formats for small-n Mann Whitney
}
\keyword{internal}
