% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_median_test.R
\name{show,dfba_median_test_out-method}
\alias{show,dfba_median_test_out-method}
\title{Formats for Bayesian Median Test Output}
\usage{
\S4method{show}{dfba_median_test_out}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_median_test_out}}}
}
\value{
No return value, called for side effect. Objects of class \code{\linkS4class{dfba_median_test_out}} are printed.
}
\description{
Formats for Bayesian Median Test Output
}
\keyword{internal}
