% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_phi.R
\name{show,dfba_phi_out-method}
\alias{show,dfba_phi_out-method}
\alias{show,dfba_phi_star_out-method}
\alias{plot,dfba_phi_out-method}
\alias{plot,dfba_phi_star_out-method}
\title{Formatted output for dfba_phi}
\usage{
\S4method{show}{dfba_phi_out}(object)

\S4method{show}{dfba_phi_star_out}(object)

\S4method{plot}{dfba_phi_out}(x, plot.prior = TRUE)

\S4method{plot}{dfba_phi_star_out}(x, plot.prior = TRUE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_phi_star_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_phi_star_out}}}

\item{plot.prior}{Show prior distribution (default = TRUE)}
}
\description{
Formatted output for dfba_phi
}
