% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_Concordance.R
\docType{class}
\name{dfba_phi_out-class}
\alias{dfba_phi_out-class}
\title{Concordance Parameter Phi}
\arguments{
\item{x}{vector of x variable values}

\item{y}{vector of y variable values}

\item{a.prior}{shape parameter a of the prior beta distribution}

\item{b.prior}{shape parameter b of the prior beta distribution}

\item{hdi.width}{Desired width of the highest density interval (HDI) of the posterior distribution (default is 95\%)}
}
\value{
A list containing the following components:

\item{Tau}{Nonparametric Tau-a correlation}

\item{sample_p}{Sample concordance proportion}

\item{nc}{Number of concordant (x, y) pairs}

\item{nd}{Number of discordant (x, y) pairs}

\item{post.median}{Median of posterior distribution on phi}

\item{post.hdi.lower}{lower limit of the HDI with width specified by hdi.width}

\item{post.hdi.upper}{upper limit of the HDI with width specified by hdi.width}
}
\description{
This function takes two vectors and shape parameters (a and b) for the prior
beta distribution (defaults are [1,1])
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
