% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_plot_beta.R
\name{dfba_plot_beta}
\alias{dfba_plot_beta}
\title{dfba_plot_beta
Plots data from dfba functions}
\usage{
dfba_plot_beta(
  a.post,
  b.post,
  a.prior = NULL,
  b.prior = NULL,
  plot.prior = FALSE
)
}
\arguments{
\item{a.post}{Shape parameter a for posterior distribution}

\item{b.post}{Shape parameter b for posterior distribution}

\item{a.prior}{Shape parameter a for prior distribution}

\item{b.prior}{Shape parameter b for prior distribution}

\item{plot.prior}{(optional) If TRUE, plots the prior distribution}
}
\value{
Plot
}
\description{
dfba_plot_beta
Plots data from dfba functions
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
