% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_plot_mann_whitney.R
\name{dfba_plot_mann_whitney}
\alias{dfba_plot_mann_whitney}
\title{dfba_plot_mann_whitney
Plots data from dfba mann whitney}
\usage{
dfba_plot_mann_whitney(x, plot.prior = TRUE)
}
\arguments{
\item{x}{A dfba_mann_whitney object}

\item{plot.prior}{(optional) If TRUE, plots the prior distribution}
}
\value{
Plot
}
\description{
dfba_plot_mann_whitney
Plots data from dfba mann whitney
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
