% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_point_BF.R
\name{show,dfba_point_BF_out-method}
\alias{show,dfba_point_BF_out-method}
\title{Formats for Point Bayes Factor}
\usage{
\S4method{show}{dfba_point_BF_out}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_point_BF_out}}}
}
\value{
No return value, called for side effect. Objects of class \code{\linkS4class{dfba_point_BF_out}} are printed.
}
\description{
Formats for Point Bayes Factor
}
\keyword{internal}
