% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_power_curve.R
\name{show,dfba_power_curve_out-method}
\alias{show,dfba_power_curve_out-method}
\alias{plot,dfba_power_curve_out-method}
\title{Formats for power curve}
\usage{
\S4method{show}{dfba_power_curve_out}(object)

\S4method{plot}{dfba_power_curve_out}(x)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_power_curve_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_power_curve_out}}}
}
\description{
Formats for power curve
}
\keyword{internal}
