% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_sign_test.R
\name{show,dfba_sign_test_out-method}
\alias{show,dfba_sign_test_out-method}
\alias{plot,dfba_sign_test_out-method}
\title{Formats for Bayesian Sign Test}
\usage{
\S4method{show}{dfba_sign_test_out}(object)

\S4method{plot}{dfba_sign_test_out}(x, plot.prior = TRUE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_sign_test_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_sign_test_out}}}

\item{plot.prior}{Show prior distribution (default = TRUE)}
}
\description{
Formats for Bayesian Sign Test
}
\keyword{internal}
