% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_sim_data.R
\name{dfba_sim_data}
\alias{dfba_sim_data}
\title{Simulated Data Generator and Inferential Comparison}
\usage{
dfba_sim_data(
  n = 20,
  a0 = 1,
  b0 = 1,
  model,
  design,
  delta,
  shape1 = 1,
  shape2 = 1,
  block.max = 0
)
}
\arguments{
\item{n}{number of values per group or paired sample}

\item{a0}{shape parameter a of the prior beta distribution}

\item{b0}{shape parameter b of the prior beta distribution}

\item{model}{hypothesized probability density function of data distribution}

\item{design}{one of "independent" or "paired" to indicate data structure}

\item{delta}{hypothesized mean difference between groups}

\item{shape1}{First shape parameter for the distribution indicated by `model` input (default is `1`)}

\item{shape2}{First shape parameter for the distribution indicated by `model` input (default is `1`)}

\item{block.max}{unclear.}
}
\value{
A list containing the following components:

\item{Tau}{Nonparametric Tau-a correlation}

\item{sample_p}{Sample concordance proportion}

\item{nc}{Number of concordant (x, y) pairs}

\item{nd}{Number of discordant (x, y) pairs}

\item{post.median}{Median of posterior distribution on phi}

\item{post.hdi.lower}{lower limit of the HDI with width specified by interval.width}

\item{post.hdi.upper}{upper limit of the HDI with width specified by interval.width}
}
\description{
Function computes the Bayesian probability that the population parameter is
greater than .5 as well as the frequentist p value via a t test. The
function is designed to be used for sampling from nine different continuous
univariate probability models. The models are the distribution on mlist
below When the design="independent", then the Bayesian analysis is in terms
of the Mann-Whitney statistic, and the t is the independent t-test.
When design="paired", then the Bayesian analysis is in terms of the Wilcoxon
signed-rank statistic, and t is paired t. The parameter delta is a
difference between the E and C conditions. The shape_vec components are
parameters for the E and C conditions. These shape_vec components are
described for each distributional model. For all models, there is a possible
blocking factor, which is a correlated values added to both the experimental
and control condition values. These add values are distributed as a uniform
on the zero to value of the blocking input. This added blocking factor has a
default of zero but positive values induces a correlation between the E and
C variates.
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
