% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_sim_data.R
\name{show,dfba_sim_data_out-method}
\alias{show,dfba_sim_data_out-method}
\alias{plot,dfba_sim_data_out-method}
\title{Format for Simulated Data Function}
\usage{
\S4method{show}{dfba_sim_data_out}(object)

\S4method{plot}{dfba_sim_data_out}(x)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_sim_data_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_sim_data_out}}}
}
\value{
No return value, called for side effect. Objects of class \code{\linkS4class{dfba_sim_data_out}} are printed.

No return value, called for side effect. Method produces a plot of class \code{\linkS4class{dfba_sim_data_out}}
}
\description{
Format for Simulated Data Function
}
\keyword{internal}
