% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfba_wilcoxon.R
\name{dfba_wilcoxon}
\alias{dfba_wilcoxon}
\title{Repeated-Measures Test (Wilcoxon Signed-Ranks Test)}
\usage{
dfba_wilcoxon(
  Y1,
  Y2,
  a0 = 1,
  b0 = 1,
  prob_interval = 0.95,
  samples = 30000,
  method = NULL
)
}
\arguments{
\item{Y1}{Numeric vector of data values representing one repeated measure}

\item{Y2}{Numeric vector of data values representing a repeated measure additional to `Y1`}

\item{a0}{shape parameter alpha of the prior beta distribution}

\item{b0}{shape parameter beta of the prior beta distribution}

\item{prob_interval}{Desired width of the highest density interval (HDI) of the posterior distribution (default is 95\%)}

\item{samples}{The number of desired Markov-Chain samples (default is 30000)}

\item{method}{(Optional) The method option is either "small" or "large". The "small" algorithm is based on a discrete Monte Carlo solution for cases where n is typically less than 20. The "large" algorithm is based on beta approximation model for the posterior distribution for the omega_E parameter. This approximation is reasonable when n > 19. Regardless of n the user can stipulate which method that they desire. When the method option is  omitted the program selects the appropriate procedure.}
}
\value{
A list containing the following components:

\item{T_plus}{Sum of the positive ranks in the pairwise comparisons}

\item{T_negative}{Sum of the negative ranks in the pairwise comparisons}
}
\description{
Wilcoxon signed rank statistic is based on paired continuous measures Y1 and
Y2. prior_vec has two components that are the shape parameters of the prior
beta distribution. The default prior is the uniform prior, but the user can
stipulate with the components of prior_vec, the desired prior distribution.
The prob_interval is value for the probability interval estimates. Samples
is the number of Monte Carlo samples used for the small n analysis. The
method option is either "small" or "large". The "small" algorithm is based
on a discrete Monte Carlo solution for cases where n is typically less than
25. The "large" algorithm is based on beta approximation model for the
posterior distribution for the phi_w parameter. This approximation is
reasonable when n > 24. Regardless of n, the user can stipulate which method
that they desire. When the method option, is omitted the program selects the
appropriate procedure.
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.
}
