% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dfba_wilcoxon_small.R
\name{show,dfba_wilcoxon_small_out-method}
\alias{show,dfba_wilcoxon_small_out-method}
\alias{plot,dfba_wilcoxon_small_out-method}
\title{Formats for small-n Wilcoxon}
\usage{
\S4method{show}{dfba_wilcoxon_small_out}(object)

\S4method{plot}{dfba_wilcoxon_small_out}(x, plot.prior = TRUE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{dfba_wilcoxon_small_out}}}

\item{x}{An object of class \code{\linkS4class{dfba_wilcoxon_small_out}}}

\item{plot.prior}{Show prior distribution (default = TRUE)}
}
\value{
No return value, called for side effect. Objects of class \code{\linkS4class{dfba_wilcoxon_small_out}} are printed.

No return value, called for side effect. Method produces a plot of class \code{\linkS4class{dfba_wilcoxon_small_out}}
}
\description{
Formats for small-n Wilcoxon
}
\keyword{internal}
