% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_W_Discrete_1.R
\name{mann_whitney_discrete_approx_1}
\alias{mann_whitney_discrete_approx_1}
\title{Mann-Whitney Discrete Function 1
This version takes as input $U_E$, $n_E$, $n_C$
the desired number of samples (default = 10000)
and the desired number of intervals (default = 200).}
\usage{
mann_whitney_discrete_approx_1(
  U_A,
  n_A,
  n_B,
  n_samples = 10000,
  n_intervals = 200
)
}
\arguments{
\item{U_A}{U statistic for the A variable}

\item{n_A}{Number of values for the A variable}

\item{n_B}{Number of values for the B variable}

\item{n_samples}{Number of desired samples}

\item{n_intervals}{Number of desired intervals for the discrete approximation}
}
\value{
omegapost

omegabar

comega

prH1 Posterior probability of the hypothesis that A > B
}
\description{
Mann-Whitney Discrete Function 1
This version takes as input $U_E$, $n_E$, $n_C$
the desired number of samples (default = 10000)
and the desired number of intervals (default = 200).
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press
}
