% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mann_Whitney_Stats.R
\name{Mann_Whitney_U_Stats}
\alias{Mann_Whitney_U_Stats}
\title{Mann-Whitney U Stats
This function takes two (2) data vectors as input
and outputs U statistics}
\usage{
Mann_Whitney_U_Stats(A, B)
}
\arguments{
\item{A}{A vector of values for variable A}

\item{B}{A vector of values for variable B}
}
\value{
UA Value of UA

UB Value of UB

U Mann-Whitney U statistic (the lesser of UA and UB)
}
\description{
Mann-Whitney U Stats
This function takes two (2) data vectors as input
and outputs U statistics
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press
}
