% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_Concordance.R
\name{Phi}
\alias{Phi}
\title{Concordance Parameter Phi
This function takes two vectors and shape parameters (a and b) for the prior
beta distribution (defaults are [1,1])}
\usage{
Phi(x, y, a.prior = 1, b.prior = 1, hdi.width = 0.95)
}
\arguments{
\item{x}{vector of x variable values}

\item{y}{vector of y variable values}
}
\value{
Tau Nonparametric Tau-a correlation

sample_p Sample concordance proportion
}
\description{
Concordance Parameter Phi
This function takes two vectors and shape parameters (a and b) for the prior
beta distribution (defaults are [1,1])
}
