% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_star.R
\name{Phi_star}
\alias{Phi_star}
\title{Parameterization-adjusted Concordance Parameter Phi*}
\usage{
Phi_star(x, y, a.prior = 1, b.prior = 1, hdi.width = 0.95, fitting.parameters)
}
\arguments{
\item{x}{Vector of values for the x variable}

\item{y}{Vector of values for the y variable}

\item{a.prior}{(Optional) Value for the shape parameter alpha (default is 1) for the prior beta distribution}

\item{b.prior}{(Optional) Value for the shape parameter beta (default is 1) for the prior beta distribution}

\item{hdi.width}{(Optional) Desired range of the highest density interval (HDI) on the posterior estimate of the phi parameter (default is 95\%)}

\item{fitting.parameters}{Number of parameters used in the predictive model}
}
\value{
\item{Tau}{sample statistic tau-a}

\item{Tau_star}{Tau-a value adjusted for the number of fitting parameters used in the predictive model}
}
\description{
This function takes two vectors,
shape parameters (a and b) for the prior
beta distribution (defaults are [1,1]),
and the number of fitting parameters used
to fit the model to the observed data.
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology
}
