% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_to_vec.R
\name{Table_to_vec}
\alias{Table_to_vec}
\title{Table-to-vector
Helper function for dfba_gamma}
\usage{
Table_to_vec(table)
}
\arguments{
\item{table}{A cross-tabulation of two variables (x and y)}
}
\value{
\item{x}{Vector of x variates grouped according to cross-tabulation in input table}

\item{y}{Vector of y variates grouped according to cross-tabulation in input table}
}
\description{
Table-to-vector
Helper function for dfba_gamma
}
\references{
Chechile, R.A. (2020). Bayesian Statistics for Experimental Scientists. Cambridge: MIT Press.

Chechile, R.A., & Barch, D.H. (2021). Distribution-free, Bayesian goodness-of-fit method for assessing similar scientific prediction equations. Journal of Mathematical Psychology.
}
